Step 2 in smart BFS traversal
Changed the way we traverse the graph when we want to create/del nodes/edges.
Instead of making the changes while performing a BFS (i.e. change the graph from
the edge/node callback) decided to use a search-and-alter approach.
Each iteration, traverse the graph and search for the pattern.  When found, make the
changes you want to make; then start the search from the beginning of the graph.
Stop when the search finds nothing.