Change the order of graph transformer processing
We need to remove nodes before looking for some of the subgraph
patterns.  For example, if looking to merge Convolution+ReLU, and
the graph contains subgraphs of the form: Conv + BatchNorm + ReLU,
we should first remove the BatchNorm nodes (fold them into the
weights) and only then look for the Conv+ReLU pattern.