Introducing "decorator transforms"
These transforms add annotations to nodes in the graph.
Now, instead of the CSV backend computing certain attributes, it
applies a set of transforms and then creates the CSV file.
The attributes are no longer available to the CSV backend, because
they can be used by anyone applying the decorators.