Experimental support for tensorflow, caffe2, and onnx networks
This is not usable at this stage.
I have added the infrastructure to pull the proto files of the
various networks, compile them, and test-use them.
Next step will be to start actually parsing them and importing graphs.

I've also downloaded Alexnet for TF and Caffe2, to use as test cases.